GibbsSampler.jl

There are different MCMC sampling algorithm is available with GibbsSampler.jl package.

MH Sampler

The MH sampler can be configured as struct like below:

GibbsSampler.MH — Type

MH

MH is a struct to choose MH MCMC sampling algorithm

Fields

n_samples : Number of iterations. Default value is 10

Example

a = MH(n_samples = 4)

adHMC Sampler

The HMC sampler can be configured as struct like below:

GibbsSampler.adHMC — Type

adHMC

adHMC is a struct to choose HMC MCMC sampling algorithm

Fields

n_samples : Number of iterations

n_adapts : Adaptation

backend : Automatic differentiation backend

Example

a = adHMC( nsamples = 10, nadapts = 5, backend = ForwardDiff )

The backend variable is used to select the automatic differentiation backend. ForwardDiff and ReversedDiff are the values to select forward and reverse differentiation, respectively.

adNUTS Sampler

The NUTS sampler can be configured as struct like below:

GibbsSampler.adNUTS — Type

adNUTS

adNUTS is a struct to choose NUTS MCMC sampling algorithm

Fields

n_samples : Number of iterations

n_adapts : Adaptation

backend : Automatic differentiation backend

Example

a = adNUTS( nsamples = 10, nadapts = 5, backend = ForwardDiff )

The backend variable is used to select the automatic differentiation backend. ForwardDiff and ReversedDiff are the values to select forward and reverse differentiation, respectively.