GibbsSampler.jl

This package helps to generate posterior samples using Gibbs sampling algorithm from a specified multivariate probability distribution when direct sampling is difficult. This Julia package supports MH and HMC based algorithms with different automatic differentiation backends.

Gibbs Sampling

GibbsSampler.gibbs — Method

function gibbs(proposal::Vector{T}, logJoint::Function; 
	sample_alg = [MH() for _ in 1:length(proposal)], 
	itr = 100, burn_in = Int(round(itr*0.2))
) where {T <: Distribution}

To generate posterior samples using Gibbs sampling algorithm

Inputs

proposal :proposal distributions as vector

logJoint :Log PDF as a function

Keyword Arguments

sample_alg :MCMC sampling for each parameter

itr :Number of iterations

burn_in :Burn in from samples

Output

chn :Generated samples